Rename grid layer classes and moved stuff to base layer
