+++ extensions/amp-story/1.0/amp-story-grid-layer.js
@@ -27,10 +27,7 @@
  */
 
 import {AmpStoryBaseLayer} from './amp-story-base-layer';
-import {StateProperty, getStoreService} from './amp-story-store-service';
-import {assertDoesNotContainDisplay, px, setStyles} from '../../../src/style';
-import {escapeCssSelectorIdent} from '../../../src/css';
-import {parseQueryString} from '../../../src/url';
+import {assertDoesNotContainDisplay, setStyles} from '../../../src/style';
 import {scopedQuerySelectorAll} from '../../../src/dom';
 
 /**
@@ -61,54 +58,12 @@ const SUPPORTED_CSS_GRID_ATTRIBUTES_SELECTOR = Object.keys(
   .join(',');
 
 /**
- * The attribute name for grid layer templates.
- * @private @const {string}
- */
-const TEMPLATE_ATTRIBUTE_NAME = 'template';
-
-/**
- * A mapping of template attribute values to CSS class names.
- * @const {!Object<string, string>}
- */
-export const GRID_LAYER_TEMPLATE_CLASS_NAMES = {
-  'fill': 'i-amphtml-story-grid-template-fill',
-  'vertical': 'i-amphtml-story-grid-template-vertical',
-  'horizontal': 'i-amphtml-story-grid-template-horizontal',
-  'thirds': 'i-amphtml-story-grid-template-thirds',
-};
-
-/**
  * Grid layer template templating system.
  */
 export class AmpStoryGridLayer extends AmpStoryBaseLayer {
   /** @param {!AmpElement} element */
   constructor(element) {
     super(element);
-
-    /** @private {?boolean} */
-    this.isPrerenderActivePage_ = null;
-
-    /** @private {?{horiz: number, vert: number}} */
-    this.aspectRatio_ = null;
-  }
-
-  /**
-   * Returns true if the page should be prerendered (for being an active page or first page)
-   * @return {boolean}
-   */
-  isPrerenderActivePage() {
-    if (this.isPrerenderActivePage_ != null) {
-      return this.isPrerenderActivePage_;
-    }
-    const hashId = parseQueryString(this.win.location.href)['page'];
-    let selector = 'amp-story-page:first-of-type';
-    if (hashId) {
-      selector += `, amp-story-page#${escapeCssSelectorIdent(hashId)}`;
-    }
-    const selectorNodes = this.win.document.querySelectorAll(selector);
-    this.isPrerenderActivePage_ =
-      selectorNodes[selectorNodes.length - 1] === this.element.parentElement;
-    return this.isPrerenderActivePage_;
   }
 
   /** @override */
@@ -117,69 +72,12 @@ export class AmpStoryGridLayer extends AmpStoryBaseLayer {
     this.applyTemplateClassName_();
     this.setOwnCssGridStyles_();
     this.setDescendentCssGridStyles_();
-    this.initializeListeners_();
+    this.initializeAspectRatioListeners();
   }
 
   /** @override */
   prerenderAllowed() {
     return this.isPrerenderActivePage();
-  }
-
-  /** @private */
-  initializeListeners_() {
-    const aspectRatio = this.element.getAttribute('aspect-ratio');
-    if (aspectRatio) {
-      const aspectRatioSplits = aspectRatio.split(':');
-      const horiz = parseInt(aspectRatioSplits[0], 10);
-      const vert = parseInt(aspectRatioSplits[1], 10);
-      if (horiz > 0 && vert > 0) {
-        this.aspectRatio_ = {horiz, vert};
-        const storeService = getStoreService(this.win);
-        storeService.subscribe(
-          StateProperty.PAGE_SIZE,
-          this.updatePageSize_.bind(this),
-          true /* callToInitialize */
-        );
-      }
-    }
-  }
-
-  /**
-   * @param {?{width: number, height: number}} pageSize
-   * @private
-   */
-  updatePageSize_(pageSize) {
-    if (!pageSize) {
-      return;
-    }
-    const {width: vw, height: vh} = pageSize;
-    const {horiz, vert} = this.aspectRatio_;
-    const width = Math.min(vw, (vh * horiz) / vert);
-    const height = Math.min(vh, (vw * vert) / horiz);
-    if (width > 0 && height > 0) {
-      this.getVsync().mutate(() => {
-        this.element.classList.add('i-amphtml-story-grid-template-aspect');
-        setStyles(this.element, {
-          '--i-amphtml-story-layer-width': px(width),
-          '--i-amphtml-story-layer-height': px(height),
-        });
-      });
-    }
-  }
-
-  /**
-   * Applies internal CSS class names for the template attribute, so that styles
-   * can use the class name instead of compound
-   * amp-story-grid-layer[template="..."] selectors, since the latter increases
-   * CSS specificity and can prevent users from being able to override styles.
-   * @private
-   */
-  applyTemplateClassName_() {
-    if (this.element.hasAttribute(TEMPLATE_ATTRIBUTE_NAME)) {
-      const templateName = this.element.getAttribute(TEMPLATE_ATTRIBUTE_NAME);
-      const templateClassName = GRID_LAYER_TEMPLATE_CLASS_NAMES[templateName];
-      this.element.classList.add(templateClassName);
-    }
   }
 
   /**
