+++ extensions/amp-story/1.0/amp-story-user-overridable.css
@@ -27,25 +27,25 @@ amp-story-grid-layer * {
   margin: 0;
 }
 
-.i-amphtml-story-grid-template-fill amp-anim img,
-.i-amphtml-story-grid-template-fill amp-img img,
-.i-amphtml-story-grid-template-fill amp-video video,
-.i-amphtml-story-grid-template-with-full-bleed-animation amp-img img {
+.i-amphtml-story-layer-template-fill amp-anim img,
+.i-amphtml-story-layer-template-fill amp-img img,
+.i-amphtml-story-layer-template-fill amp-video video,
+.i-amphtml-story-layer-template-with-full-bleed-animation amp-img img {
   object-fit: cover;
 }
 
-.i-amphtml-story-grid-template-vertical {
+.i-amphtml-story-layer-template-vertical {
   align-content: start;
   grid-gap: 16px;
   justify-content: stretch;
   justify-items: start;
 }
 
-.i-amphtml-story-grid-template-vertical > * {
+.i-amphtml-story-layer-template-vertical > * {
   width: 100%;
 }
 
-.i-amphtml-story-grid-template-horizontal {
+.i-amphtml-story-layer-template-horizontal {
   align-content: stretch;
   align-items: start;
   grid-gap: 16px;
