+++ extensions/amp-story/1.0/test/test-amp-story-grid-layer.js
@@ -91,7 +91,7 @@ describes.realWin('amp-story-grid-layer', {amp: true}, (env) => {
 
     storeService.dispatch(Action.SET_PAGE_SIZE, {width: 1000, height: 1000});
 
-    expect(gridLayerEl).to.have.class('i-amphtml-story-grid-template-aspect');
+    expect(gridLayerEl).to.have.class('i-amphtml-story-layer-template-aspect');
     expect(
       parseInt(
         gridLayerEl.style.getPropertyValue('--i-amphtml-story-layer-width'),
@@ -112,7 +112,7 @@ describes.realWin('amp-story-grid-layer', {amp: true}, (env) => {
 
     storeService.dispatch(Action.SET_PAGE_SIZE, {width: 1000, height: 1000});
 
-    expect(gridLayerEl).to.have.class('i-amphtml-story-grid-template-aspect');
+    expect(gridLayerEl).to.have.class('i-amphtml-story-layer-template-aspect');
     expect(
       parseInt(
         gridLayerEl.style.getPropertyValue('--i-amphtml-story-layer-width'),
